[Feat] #292 - 재구독뷰 구현 완료

+++ YELLO-iOS/YELLO-iOS/Network/Purchase/DTO/Response/PurchaseSubscibeNeedResponseDTO.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 struct PurchaseSubscibeNeedResponseDTO: Codable {
-    let id: Int64
+    let id: Int
     let subscribe: String
     let expiredDate: String
 }
